ver4.2

+++ main_control_local.py
@@ -3,26 +3,39 @@ import datetime
 import os 
 import subprocess
 import glob
+import sys
+
+# import numpy as np
+# import pandas as pd
+# import binascii
+# import csv
+# from astropy.io import fits
 
 # memory test
-import psutil
-def print_memory():
-    process = psutil.Process(os.getpid())
-    mem = process.memory_info().rss / (1024 * 1024)  # Memory in MB
-    print(f"[Memory] {mem:.2f} MB")
+# import psutil
+# def print_memory():
+#     process = psutil.Process(os.getpid())
+#     mem = process.memory_info().rss / (1024 * 1024)  # Memory in MB
+#     print(f"[Memory] {mem:.2f} MB")
+
+# Check for new files every x seconds
+check_time = 5 
+csv_header = 'Filename,Type,Start_Packet_number,End_Packet_number,Incompleteness\n'
 
 raw_data_folder = "./raw_data/"
 req_data_folder = "./requested_data/"
-img_data_folder = "./optical/"
 log_folder = "./log/"
 archive_raw_folder = "./archive/raw_data/"
 archive_req_folder = "./archive/requested_data/"
 os.makedirs(log_folder, exist_ok=True)
 os.makedirs(raw_data_folder, exist_ok=True)
 os.makedirs(req_data_folder, exist_ok=True)
-os.makedirs(img_data_folder, exist_ok=True)
 os.makedirs('./report/', exist_ok=True)
-os.makedirs('./img/', exist_ok=True)
+os.makedirs('./fits/', exist_ok=True)
+os.makedirs('./csv/', exist_ok=True)
+os.makedirs('./txt/', exist_ok=True)
+os.makedirs('./log/', exist_ok=True)
+os.makedirs('./jpg/', exist_ok=True)
 os.makedirs('./tmp/', exist_ok=True)
 os.makedirs('./cmd/', exist_ok=True)
 os.makedirs('./cmd/list/', exist_ok=True)
@@ -32,19 +45,19 @@ os.makedirs(archive_req_folder, exist_ok=True)
 time_now = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
 nfiles = len(glob.glob(log_folder + "*.log"))
 log_file = log_folder + f"log_{nfiles}_{time_now}.log"
-os.system(f"touch {log_file}")
+subprocess.run(['touch', log_file])
 
-fout_name_cpl = './report/final_check.csv'
-with open(fout_name_cpl, 'w') as f:
-    f.write('Filename,Type,Start_Packet_number,End_Packet_number,Incompleteness(100*missing/16621)\n')
+cmd_gen_files = './report/un_gen.csv'
+check_file = './report/final_check.csv'
 final_report = './report/report.csv'
-with open(final_report, 'w') as f:
-    f.write('Filename,Type,Start_Packet_number,End_Packet_number,Incompleteness(100*missing/16621)\n')
+if os.path.exists(final_report):
+    pass
+else:
+    with open(final_report, 'w') as f:
+        f.write(csv_header)
 
 while True:
-    
-    # test memory
-    print_memory()
+    # print_memory()
     
     processed_raw_files = set()
     processed_req_files = set()
@@ -67,15 +80,13 @@ while True:
     current_img_files = {f for f in os.listdir(img_data_folder) if os.path.isfile(os.path.join(img_data_folder, f))}
     new_img_files = current_img_files - processed_img_files
 
+    # call check_data.py
     for file in sorted(new_raw_files):  # Process in order
         file_path = os.path.join(raw_data_folder, file)
         with open(log_file, "a") as f:
             f.write(f"Checking {file}\n")
         try:
             subprocess.run(["python3", "./check_data.py", file_path], check=True)
-            # test memory
-            print_memory()
-            
             with open(log_file, "a") as f:
                 f.write(f"Finish checking {file}\n")
             processed_raw_files.add(file)
@@ -83,14 +94,12 @@ while True:
             with open(log_file, "a") as f:
                 f.write(f"Error for checking {file_path}: {e}\n")
                 f.write(f"Delete {file}, request again.\n")
-            os.system(f'rm {raw_data_folder}{file}')
+            subprocess.run(['rm', file_path])
             continue
             
-    # call cmd_gen.py
-    os.system(
-        f"python3 ./cmd_gen.py"
-    )
+    # print_memory()
     
+    # call combine.py
     for file in sorted(new_req_files):  # Process in order
         file_path = os.path.join(req_data_folder, file)
         with open(log_file, "a") as f:
@@ -104,11 +113,15 @@ while True:
             with open(log_file, "a") as f:
                 f.write(f"Error for extracting {file_path}: {e}\n")
                 f.write(f"Delete {file}.\n")
-            os.system(f'rm {new_req_files}{file}')
+            subprocess.run(['rm', file_path])
             continue
-        
+    
+    # print_memory()
+    
+    # call read_bin.py
     for file in sorted(new_img_files):  # Process in order, file = opt_frame_n_Fxxx.bin
         file_path = os.path.join(img_data_folder, file)
+        file_originame = file.split('_')[-1] # Fxxx.bin
         with open(log_file, "a") as f:
             f.write(f"Reading {file}\n")
         try:
@@ -116,58 +129,78 @@ while True:
             with open(log_file, "a") as f:
                 f.write(f"Finished compiling image from {file}\n")
             processed_img_files.add(file)
-            # complete file, cut the report in final_check.csv to report.csv
-            with open('./report/final_check.csv', 'r') as f1:
-                lines = f1.readlines()
-            # Open report.csv to append the matched line
-            with open('./report/report.csv', 'a') as f2:
+            # Read the report of file_originame from check_file and append it to final_report
+            with open(check_file, 'r') as f1:
+                lines = f1.readlines()[1:]  # Skip the header
+            with open(final_report, 'a') as f2:
                 new_lines = []
                 for line in lines:
-                    if line.split(',')[0] == file.split('_')[-1]:
-                        f2.write(line)  # Append to report.csv
+                    if line.split(',')[0] == file_originame:
+                        f2.write(line)  # Append to final_report
                     else:
-                        new_lines.append(line)  # Keep other lines
-            # Overwrite final_check.csv with remaining lines
-            with open('./report/final_check.csv', 'w') as f1:
-                f1.writelines(new_lines)
+                        new_lines.append(line)  # Keep other lines (original lines in check_file)
                 
+            # Overwrite check_file with remaining lines
+            with open(check_file, 'w') as f:
+                f.write(csv_header)  # Write header
+                f.writelines(new_lines)
+                
+            # report the completed file to cmd_gen_files. Final confirmation.
+            with open(cmd_gen_files, 'a') as f3:
+                f3.write(f'{file_originame},OK,0,0,0\n')
+            
+        # file failed to compile. Copy report from check_file to final_report
         except subprocess.CalledProcessError as e:
             with open(log_file, "a") as f:
                 f.write(f"Error for reading {file_path}: {e}\n")
-            # incomplete file, request again
-            with open('./report/final_check.csv', 'r') as f1:
+            # corrupted file, request again
+            with open(check_file, 'r') as f1:
                 lines = f1.readlines()
-            # Open report.csv to append the matched line
-            with open('./report/report.csv', 'a') as f2:
+            with open(final_report, 'a') as f2:
                 new_lines = []
                 for line in lines:
                     if line.split(',')[0] == file.split('_')[-1]:
                         f2.write(f'{file},Error,65535,65535,100\n')  # Append to report.csv
                     else:
-                        new_lines.append(line)  # Keep other lines
-            # Overwrite final_check.csv with remaining lines
-            with open('./report/final_check.csv', 'w') as f1:
-                f1.writelines(new_lines)
+                        new_lines.append(line)  # Keep other lines (original lines in check_file)
+            
+            # Overwrite check_file with remaining lines
+            with open(check_file, 'w') as f:
+                f.write(csv_header)  # Write header
+                f.writelines(new_lines)
+                
+            # report the corrupted file to cmd_gen_files. 
+            with open(cmd_gen_files, 'a') as f3:
+                f3.write(f'{file_originame},Error,65535,65535,100\n')
             
             with open(log_file, "a") as f:
                 f.write(f"Delete {file}, request again.\n")
-            os.system(f'rm {img_data_folder}{file}')
+            subprocess.run(['rm', file_path])
                 
             continue
 
-    time.sleep(3)  # Check for new files every x seconds
+
+    # call cmd_gen.py
+    subprocess.run(['python3', './cmd_gen.py'])
     
+    # print_memory()
+    
+    time.sleep(check_time)
+    
+    # clear the processed files
     for file in processed_raw_files:
         with open(log_file, "a") as f:
             f.write(f"Move {file} to archive\n")
-        os.system(f'mv {raw_data_folder}{file} {archive_raw_folder}{file}')
+        subprocess.run(['mv', f'{raw_data_folder}{file}', f'{archive_raw_folder}{file}'])
     for file in processed_req_files:
         with open(log_file, "a") as f:
             f.write(f"Move {file} to archive\n")
-        os.system(f'mv {req_data_folder}{file} {archive_req_folder}{file}')
+        subprocess.run(['mv', f'{req_data_folder}{file}', f'{archive_req_folder}{file}'])
     for file in processed_img_files:
         with open(log_file, "a") as f:
             f.write(f"Delete {file}\n")
-        os.system(f'rm {img_data_folder}{file}')
+        subprocess.run(['rm', f'{img_data_folder}{file}'])
     # print(f'files: {current_files}')
+    
+    # print_memory()
